Fix default quota value for new stores
- Make default quota value for new stores created via BnP to be configurable
- -1 for default quota value indicates no quota restriction
- fix unit tests

fix tests

minor fix

Fix bug

- delete quota call for, setQuoataForNode() now deletes quota from the quota store instead of the
  actual store name

+++ src/java/voldemort/tools/admin/command/AdminCommandQuota.java
@@ -27,7 +27,6 @@ import java.util.Set;
 import joptsimple.OptionParser;
 import joptsimple.OptionSet;
 import voldemort.VoldemortException;
-import voldemort.client.SystemStoreClient;
 import voldemort.client.protocol.admin.AdminClient;
 import voldemort.store.quota.QuotaType;
 import voldemort.store.quota.QuotaUtils;
@@ -574,7 +573,7 @@ public class AdminCommandQuota extends AbstractAdminCommand {
                                 adminClient.quotaMgmtOps.setQuotaForNode(storeName,
                                                                          QuotaType.valueOf((String) entry.getKey()),
                                                                          nodeId,
-                                                                         Integer.parseInt((String) entry.getValue()));
+                                                                         Long.parseLong((String) entry.getValue()));
                             }
                         }
                     }
